requirements.txt
django>=3.0,<4.0
djangorestframework==3.14.0
Markdown==3.5.2
